Skip a date-time test when we don't know what result it should give
The operator_eqeq(data13) test expected the local-time epoch and UTC
epoch to agree precisely if the localTimeType set by the test's
constructor says local time is UTC; however, when the local zone is
*sometimes* ahead of (or behind) UTC, due to DST, localTimeType is
duly set to indicate that, which doesn't preclude the zone agreeing
with UTC at the epoch.  This indeed happens for Europe/London, which
agrees on the epoch but was ahead a few months later.  So we can't
determine what outcome to expect based solely on localTimeType,
although we can be sure of a match when local time is UTC.  So skip
this test when local time isn't UTC (and document what's going on a
bit better).

Task-number: QTBUG-65435
Change-Id: Id9b8aa0402f2a2b410e0234f6eca4ab0d1010bc4
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>